move CheckForRoot and CheckCert classes into EnvironmentChecker util class, add method to check if app is debugabble
